# Remove symbols map from TSDB head