This saves memory, effort and locking.

Since every symbol is also added to postings, `Symbols()` can be implemented there instead.

This now has to build a map for deduplication, but `Symbols()` is only called for compaction, and `gc()` used to rebuild the symbols map after every compaction so not an additional cost.

Benchmark:
```
name            old time/op    new time/op    delta
CreateSeries-8    2.02µs ± 8%    1.50µs ± 1%  -25.78%  (p=0.016 n=5+4)

name            old alloc/op   new alloc/op   delta
CreateSeries-8    1.19kB ± 5%    1.08kB ± 5%   -8.85%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
CreateSeries-8      5.00 ± 0%      5.00 ± 0%     ~     (all equal)
```