@bboreham 2021-09-07
Looking at the prombench dashboard, everything is very close between the two versions, so at least I haven't made things worse.

On the other hand it doesn't show up as any better, but in environments (e.g. Cortex) where many more appenders are used in parallel I am confident this will make an improvement.